backend/agents/parser_agent.py
# agents/parser_agent.py

from crewai import Agent
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Parser Agent: Extracts and interprets user product requirements
parser_agent = Agent(
    role="Product Request Intelligence Parser",
    goal=(
        "Analyze user requests to extract detailed product requirements such as "
        "category, specifications, budget, and preferences. Generate optimized "
        "search strategies tailored to the request."
    ),
    backstory=(
        "An AI specialist in natural language understanding with expertise in "
        "consumer electronics. Skilled at interpreting complex product needs, "
        "recognizing use cases like gaming, productivity, or creative work, "
        "and identifying both explicit and implicit requirements. "
        "Ensures that extracted details are accurate and context-aware."
    ),
    verbose=True,
    allow_delegation=False
)
