backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1

# Database and Supabase
supabase==2.0.0

# AI and LLM frameworks
crewai==0.70.0
langchain==0.3.26
langchain-core==0.3.74
langchain-community==0.3.26
langchain-google-genai==2.0.0
openai==1.90.0

# HTTP and async requests
httpx==0.28.1
aiohttp==3.11.13
requests==2.32.4

# Web scraping
playwright==1.55.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Data processing and validation
pydantic==2.10.6
pydantic-settings==2.8.1
pandas==2.1.4
numpy==1.26.4

# Utility libraries
python-multipart==0.0.20
typing-extensions==4.12.2
tenacity==8.2.3

# Development and testing
pytest==8.0.0
pytest-asyncio==0.24.0